Extract function-call scope effects into FuncCallScopeEffectsHelper
